First attempt at NUMA support.
To disable NUMA support while compiling, add numa=no to the make command.
NUMA support on Linux requires libnuma and corresponding headers to be present.

If Cfish does not detect more than 1 node at startup, NUMA support is disabled.
Otherwise, a NUMA UCI option is present which defaults to "all".
On Windows, valid values for NUMA are "off", "on" and "all". There is
no difference between "on" and "all".
On Linux, the NUMA option may additionally be set, for example,  to "0-2,5"
to indicate that search threads may be launched on nodes 0, 1, 2 and 5.
The syntax is the same as for the numactl command (see man numactl).
Only nodes present in the system may be specified (otherwise the value set
is ignored).

NUMA support binds threads to nodes and allocates thread-specific data
at the node to which the thread is bound. In addition, on Linux the
TT table is allocated interleaved across all nodes (or the subset of nodes
specified in the NUMA option).

Threads are mapped to nodes as follows.
- First, the first available nodes gets as many threads as it has physical
  cores. Then the second available node is filled, etc.
- Further threads are assigned round-robin to the available nodes.